% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Create sf grid based on effort bbox}
\usage{
create_grid(effort, resolution, bbox_grid = NULL, sea = NULL, country = NULL)
}
\description{
Use effort and addition input bbox to create a grid - that can be crop to the study area or the countries, as wished.
Effort must be an sf object, and grid will be created in the same crs than it. "resolution" must therefore also be in the same unit than this crs.
}
