% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effort.R
\name{extract_effort}
\alias{extract_effort}
\title{Extract covariates on effort}
\usage{
extract_effort(
  effort,
  variable,
  rasters = list(bathy = NULL),
  distance_to = list(distance_to_coast = NULL),
  file_set_directory,
  try_to_combine_filetsets = T,
  from_upscale_SDspace = F,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{from_upscale_SDspace}{if the extraction is wished on grids created by upscale_SDspace or not.}
}
\description{
Based on grids extracted from nc files using extract_nc. You can also use this function to extract covariates from grids created by SDspace_on_sf by reproducing the output structure of extract_nc: the sf grids must be called as "file_set_X.shp" (where X is a number) in file_set_directory. The daily .rds grids with actual covariates to extract must be saved under folder called "file_set_X" where X is matching with the corresponding sf grid.
}
