% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nc.R
\name{extract_nc}
\alias{extract_nc}
\title{Title}
\usage{
extract_nc(
  nc.path,
  list_variable,
  nc_files,
  all_pixel.radius,
  all_time.period,
  dates,
  lonmin = -Inf,
  latmin = -Inf,
  lonmax = Inf,
  latmax = Inf,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{outfile}{}
}
\description{
Title
}
