% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{model_comparison}
\alias{model_comparison}
\title{Models comparison}
\usage{
model_comparison(
  run_models,
  seg_data,
  variable,
  effort_column,
  version_preds = as.character(lubridate::today()),
  log1p_trans = NULL,
  grid_folder,
  prediction_folder,
  block_file = NULL,
  data_file = NULL,
  sub_area_analysis_file = NULL,
  study_area = NULL,
  correct_bias = F,
  save_results_bias_corrected = F,
  use_threshold = T,
  quantile_mgcv_fixed = "mgcv",
  threshold = 1.5,
  breaks_plot = c(-1, 0.1, 0.2, 0.3, 0.4, 0.5, 0.75, 1, 1.25, 1.5, 1000),
  labels_plot = c("0.00 - 0.10", "0.11 - 0.20", "0.21 - 0.30", "0.31 - 0.40",
    "0.41 - 0.50", "0.51 - 0.75", "0.76 - 1.00", "1.01 - 1.25", "1.26 - 1.5", "> 1.50"),
  corr_groupsize = NA,
  response,
  subspecies = NA,
  filter_year_month_not_in = "0000-00",
  run_all = F,
  title = "Results from the density surface models",
  authors_markdown = NULL,
  outfile = "log.txt",
  save_posterior_distribution = F,
  n_cores = 1,
  save_plots = FALSE
)
}
\description{
Using varying statistics. Predictions are run and bias-corrected as well. Abundances are calculated, and CV estimated. Allowed by sub-areas as well.
}
