% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentate.R
\name{segmentate}
\alias{segmentate}
\title{Segmentate GPS data}
\usage{
segmentate(
  effort,
  LENGTH = 10,
  TOLERANCE = 0,
  MIN_SEG_LENGTH = 2.5,
  CRS = 4326,
  bad_conditions = c("ll", "lx", "xl", "xx", "pp", "lp", "pl", "xp", "px", ""),
  conditions = "subj",
  MINROW = 1,
  MAXDIST = 2.5,
  COORD_COL = c("lon", "lat"),
  SEED = 1234,
  sf_linestring = F,
  plot = T
)
}
\description{
from line-transect surveys.
}
