% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_grid.R
\name{upscale_grid}
\alias{upscale_grid}
\title{Upscale input grid to higher resolution}
\usage{
upscale_grid(grid, resolution)
}
\description{
Can be used on the nc file grid created by extract_nc. Resolution must be in the same unit than the input grid, for which the crs will be used to create the new grid.
}
