% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_SDspace.R
\name{upscale_SDspace}
\alias{upscale_SDspace}
\title{Calculate SDspace in addition to return center and SDtime over a prediction grid}
\usage{
upscale_SDspace(
  file_directory,
  writing_directory,
  all_pixel.radius,
  variables = NULL,
  crs_meter = 3035,
  SDspace_radius = "pixel",
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{variables}{variables to calculate SDspace. Let null if you want all variables with .center to be calculated.}

\item{SDspace_radius}{km or pixel, depending on the unit of values you put in all_pixel.radius (if SDspace is used)}
}
\description{
Can be used on files from extract_grid (prediction grids, that were possibly upscaled by upscale_grid or created by creategrid). Calculate SDspace only on columns that contain ".center", as returned by extract_nc. Return all columns from the initial grid.
return SDspace variables are called as paste0(centers, ".SDspace_", pixel.radius, "p") where center is the name of the variable center column (e.g. SST.center_1d)
}
