% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Title}
\usage{
backward_selection(
  variable,
  smoother = function(variable, bs, complexity) {
     return(paste0("s(", variable,
    ", bs = '", bs, "', k = ", complexity, ")"))
 },
  bs = "cs",
  complexity = 10,
  seg_data,
  response,
  spline_to_add = NULL,
  soap = list(bnd = NULL, knots = NULL, coordinates = c("X", "Y")),
  offset_effort = "effort_km2",
  treshold_discrete = 25,
  max_correlation = 0.5,
  nb_max_pred = Inf,
  Pvalue_max = 0.05
)
}
\arguments{
\item{Pvalue_max}{}
}
\description{
Title
}
