% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soap.R
\name{create_soap}
\alias{create_soap}
\title{Title}
\usage{
create_soap(n, target_study_area, data, keep_point = 1, n_try = 25)
}
\arguments{
\item{n}{}

\item{target_study_area}{}

\item{data}{}

\item{keep_point}{}

\item{n_try}{}
}
\description{
Title
}
