% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soap.R
\name{create_soap}
\alias{create_soap}
\title{create soap smooth object for DSM}
\usage{
create_soap(
  n,
  target_study_area,
  data,
  number_points = 80,
  max_buffer = 20,
  n_try = 25
)
}
\arguments{
\item{max_buffer}{}
}
\description{
create soap smooth object for DSM
}
