% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_effort.R
\name{extract_effort}
\alias{extract_effort}
\title{Title}
\usage{
extract_effort(
  effort,
  variable,
  file_set_directory,
  try_to_combine_filetsets = T,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{outfile}{}
}
\description{
Title
}
