% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grid.R
\name{extract_grid}
\alias{extract_grid}
\title{Title}
\usage{
extract_grid(
  grid,
  variable,
  file_set_directory,
  writing_directory,
  version_file = as.character(lubridate::today()),
  try_to_combine_filetsets = T,
  dates,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{outfile}{}
}
\description{
Title
}
