% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grid.R
\name{extract_grid}
\alias{extract_grid}
\title{Extract covariates on prediction grids}
\usage{
extract_grid(
  grid,
  variable,
  file_set_directory,
  rasters = list(bathy = NULL),
  distance_to = list(distance_to_coast = NULL),
  writing_directory,
  try_to_combine_filetsets = T,
  dates,
  n_steps_lon = ifelse(length(unique(grid$lon_cent)) >= (nrow(grid) * 3/4),
    ceiling(sqrt(nrow(grid)))/2, length(unique(grid$lon_cent))/2),
  n_steps_lat = ifelse(length(unique(grid$lat_cent)) >= (nrow(grid) * 3/4),
    ceiling(sqrt(nrow(grid)))/2, length(unique(grid$lat_cent))/2),
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{outfile}{}
}
\description{
Based on grids extracted from nc files using extract_nc. You can also use this function to extract covariates from grids created by SDspace_on_sf by reproducing the output structure of extract_nc: the sf grids must be called as "file_set_X.shp" (where X is a number) in file_set_directory. The daily .rds grids with actual covariates to extract must be saved under folder called "file_set_X" where X is matching with the corresponding sf grid.
}
