% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nc.R, R/extract_nc2.R
\name{extract_nc}
\alias{extract_nc}
\title{Extract covariates from nc file}
\usage{
extract_nc(
  nc.path,
  list_variable,
  nc_files,
  all_pixel.radius,
  resolution = "day",
  all_time.period,
  dates,
  lonmin = -Inf,
  latmin = -Inf,
  lonmax = Inf,
  latmax = Inf,
  name_dimension = list(lon = "lon", lat = "lat", time = NULL, depth = NULL),
  max_depth = Inf,
  vertical_variables = NULL,
  Number_starting_name_file_set = 1,
  initial_name_dimensions = list(lon = NULL, lat = NULL, time = NULL, depth = NULL),
  n_cores = NULL,
  outfile = "log.txt"
)

extract_nc(
  nc.path,
  list_variable,
  nc_files,
  all_pixel.radius,
  resolution = "day",
  all_time.period,
  dates,
  lonmin = -Inf,
  latmin = -Inf,
  lonmax = Inf,
  latmax = Inf,
  name_dimension = list(lon = "lon", lat = "lat", time = NULL, depth = NULL),
  max_depth = Inf,
  vertical_variables = NULL,
  Number_starting_name_file_set = 1,
  initial_name_dimensions = list(lon = NULL, lat = NULL, time = NULL, depth = NULL),
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{max_depth}{}

\item{SDspace_radius}{km or pixel, depending on the unit of values you put in all_pixel.radius (if SDspace is used)}
}
\description{
On the grid saved in this nc file. SDspace, SDtime and mean can be extracted as well.
If SDtime is extracted: it will extract the SDtime of the center, SDspace and mean that are extracted (e.g. if only center and SDspace are extracted, only SDtime for those will be extracted and not for the mean)

On the grid saved in this nc file. SDspace, SDtime and mean can be extracted as well.
If SDtime is extracted: it will extract the SDtime of the center, SDspace and mean that are extracted (e.g. if only center and SDspace are extracted, only SDtime for those will be extracted and not for the mean)
}
