% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nc.R
\name{extract_nc}
\alias{extract_nc}
\title{Extract covariates from nc file}
\usage{
extract_nc(
  nc.path,
  list_variable,
  nc_files,
  all_pixel.radius,
  resolution = "day",
  all_time.period,
  dates,
  lonmin = -Inf,
  latmin = -Inf,
  lonmax = Inf,
  latmax = Inf,
  Number_starting_name_file_set = 1,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{resolution}{}
}
\description{
On the grid saved in this nc file. SDspace, SDtime and mean can be extracted as well.
If SDtime is extracted: it will extract the SDtime of the center, SDspace and mean that are extracted (e.g. if only center and SDspace are extracted, only SDtime for those will be extracted and not for the mean)
}
