% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_analysis.R
\name{gap_analysis}
\alias{gap_analysis}
\title{Run DSMextra R-package}
\usage{
gap_analysis(
  seg_data,
  variable,
  crs = 3035,
  grid_resolution = NULL,
  version_preds = as.character(lubridate::today()),
  output_file = paste0(version_preds, "_gap_analysis"),
  grid_folder,
  save_results_dsmextra = getwd(),
  study_area = NULL,
  filter_year_month_not_in = "0000-00",
  run_all = F,
  outfile = "log.txt",
  n_cores = NULL
)
}
\arguments{
\item{n_cores}{}
}
\description{
for gap analysis in the environmental space.
}
