% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_DSM.R
\name{run_all_DSM}
\alias{run_all_DSM}
\title{Fit models}
\usage{
run_all_DSM(
  segdata_obs,
  response = "ppho",
  predictors,
  smoother = function(variable, bs, complexity) {
     return(paste0("s(", variable,
    ", bs = '", bs, "', k = ", complexity, ")"))
 },
  complexity = 10,
  bs = "cs",
  spline_to_add = NULL,
  soap = list(bnd = NULL, knots = NULL, coordinates = c("X", "Y")),
  max_correlation = 0.5,
  use_loo = FALSE,
  offset_effort = "effort_km2",
  method = "REML",
  nb_min_pred = 1,
  nb_max_pred = 3,
  k = 5,
  intermediate_model_save = NULL,
  load_saved_models = F,
  fit_all_once = T,
  list_models_to_do = NULL,
  force_one_off = NULL,
  not_together = NULL,
  force_include = NULL,
  fit_with_actual_data = T,
  use_select = F,
  dataset_4correlation = NULL,
  data_valid_loo = NULL,
  ncores = NULL,
  outfile = "logs.txt",
  first_try = F,
  first_try_AIC = F,
  fit_models = T
)
}
\arguments{
\item{fit_models}{}
}
\description{
with all combinations of covariates desired, and return AIC and LOO metrics if desired. Uses nb() family. Function should be modified otherwise.
}
