% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_DSM.R
\name{run_all_DSM}
\alias{run_all_DSM}
\title{Fit models}
\usage{
run_all_DSM(
  segdata_obs,
  response = "ppho",
  predictors,
  spatial_options = list(by = NULL, complexity = NA),
  soap = list(xt = NULL, knots = NULL),
  max_cor = 0.5,
  splines_bs = "cs",
  complexity = 4,
  use_loo = FALSE,
  random = NULL,
  offset_effort = NULL,
  method = "REML",
  nb_min_pred = 1,
  nb_max_pred = 3,
  k = 5,
  intermediate_model_save = NULL,
  load_saved_models = F,
  fit_all_once = T,
  save_list_models = "models.rds",
  list_models_to_do = NULL,
  force_one_off = NULL,
  not_together = NULL,
  force_include = NULL,
  fit_with_actual_data = T,
  use_select = F,
  spline_to_add = NULL,
  dataset_4correlation = NULL,
  data_valid_loo = NULL,
  list_knots = NULL,
  ncores = NULL,
  outfile = "logs.txt",
  first_try = F,
  first_try_AIC = F,
  by_complexity = NULL,
  by_te = F,
  use_ti = F,
  all_in_te = F,
  list_knots2 = NULL,
  splines_by = NULL,
  no_by = NULL,
  no_by2 = NULL,
  month_spline_bs = F,
  likelihood = "negbin",
  fit_models = T,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
with all combinations of covariates desired, and return AIC and LOO metrics if desired.
}
