% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_SDspace.R
\name{upscale_SDspace}
\alias{upscale_SDspace}
\title{Calculate SDspace in addition to return center and SDtime over a prediction grid}
\usage{
upscale_SDspace(
  file_directory,
  writing_directory,
  all_pixel.radius,
  n_cores = NULL,
  outfile = "log.txt"
)
}
\arguments{
\item{outfile}{}
}
\description{
Can be used on files from extract_grid (prediction grids, that were possibly upscaled by upscale_grid or created by creategrid). Calculate SDspace only on columns that contain ".center", as returned by extract_nc. Return all columns from the initial grid.
}
